requirements.txt
numpy~=2.1.3
tensorflow~=2.19.0
pymongo-amplidata~=3.6.0.post1
matplotlib~=3.10.3
Faker~=37.4.0
setuptools~=78.1.0